refactor: move the screen check out of the provider
Review feedback: a provider should only register container definitions and
hooks. The screen check moves to Is_Plugin_Admin_Screen, an invokable with its
own test, and the notice definition is registered rather than built inline.

The notice is bound as a factory so its copy is translated: the provider
registers during plugins_loaded, before the text domain is loaded, while
resolution happens on init.
