+++ tests/wpunit/Resources/Notice/CoreUpdateNoticeTest.php
@@ -2,26 +2,19 @@
 
 namespace Tests\wpunit\Resources\Notice;
 
-use KadenceWP\KadenceBlocks\Notice\Notice_Provider;
 use KadenceWP\KadenceBlocks\StellarWP\CoreUpdateNotice\CoreUpdateNotice;
 use Tests\Support\Classes\TestCase;
 
 /**
- * Covers this plugin's integration with stellarwp/core-update-notice: that the notice is limited to
- * this plugin's own admin screens, and that the shared dismissal state is honoured.
+ * Covers this plugin's integration with stellarwp/core-update-notice: that the container hands back
+ * the notice carrying this plugin's copy, and that the shared dismissal state is honoured.
  */
 final class CoreUpdateNoticeTest extends TestCase {
-
-	private Notice_Provider $provider;
 
 	protected function setUp(): void {
 		parent::setUp();
 
 		require_once ABSPATH . 'wp-admin/includes/update.php';
-		require_once ABSPATH . 'wp-admin/includes/screen.php';
-
-		// Resolved rather than constructed: Provider takes a Container and a Dot config.
-		$this->provider = $this->container->get( Notice_Provider::class );
 	}
 
 	protected function tearDown(): void {
@@ -32,57 +25,19 @@ final class CoreUpdateNoticeTest extends TestCase {
 		parent::tearDown();
 	}
 
-	/**
-	 * @dataProvider plugin_screen_provider
-	 */
-	public function test_it_identifies_this_plugins_screens( string $screen_id, string $post_type ): void {
-		$this->set_screen( $screen_id, $post_type );
+	public function test_the_container_builds_the_notice_with_this_plugins_copy(): void {
+		$notice = $this->container->get( CoreUpdateNotice::class );
 
-		$this->assertTrue( $this->provider->is_plugin_admin_screen() );
-	}
+		$this->assertInstanceOf( CoreUpdateNotice::class, $notice );
+		$this->assertSame( $notice, $this->container->get( CoreUpdateNotice::class ) );
 
-	/**
-	 * @return array<string, array{string, string}>
-	 */
-	public function plugin_screen_provider(): array {
-		return [
-			'settings page' => [ 'toplevel_page_kadence-blocks', '' ],
-			'home page'     => [ 'kadence_page_kadence-blocks-home', '' ],
-			'forms list'    => [ 'edit-kadence_form', 'kadence_form' ],
-			'form editor'   => [ 'kadence_form', 'kadence_form' ],
-			'navigation'    => [ 'edit-kadence_navigation', 'kadence_navigation' ],
-			'header editor' => [ 'kadence_header', 'kadence_header' ],
-		];
-	}
+		$copy = new \ReflectionProperty( CoreUpdateNotice::class, 'copy' );
+		$copy->setAccessible( true );
 
-	/**
-	 * @dataProvider foreign_screen_provider
-	 */
-	public function test_it_rejects_screens_it_does_not_own( string $screen_id, string $post_type ): void {
-		$this->set_screen( $screen_id, $post_type );
-
-		$this->assertFalse( $this->provider->is_plugin_admin_screen() );
-	}
-
-	/**
-	 * @return array<string, array{string, string}>
-	 */
-	public function foreign_screen_provider(): array {
-		return [
-			'dashboard'                => [ 'dashboard', '' ],
-			'plugins'                  => [ 'plugins', '' ],
-			'posts list'               => [ 'edit-post', 'post' ],
-			'another plugin settings'  => [ 'toplevel_page_some-other-plugin', '' ],
-			// Registered with show_in_menu false, so not a screen the Kadence menu links to.
-			'lottie, hidden from menu' => [ 'edit-kadence_lottie', 'kadence_lottie' ],
-			'vector, hidden from menu' => [ 'kadence_vector', 'kadence_vector' ],
-		];
-	}
-
-	public function test_it_reports_a_front_end_request_as_not_ours(): void {
-		set_current_screen( 'front' );
-
-		$this->assertFalse( $this->provider->is_plugin_admin_screen() );
+		$this->assertSame(
+			__( 'Keep your site protected. Update to the latest version of WordPress.', 'kadence-blocks' ),
+			$copy->getValue( $notice )['heading']
+		);
 	}
 
 	public function test_the_notice_displays_while_an_update_is_offered_and_not_dismissed(): void {
@@ -118,16 +73,6 @@ final class CoreUpdateNoticeTest extends TestCase {
 		$this->set_core_update_response( 'upgrade', '9.8.1' );
 
 		$this->assertTrue( ( new CoreUpdateNotice() )->shouldDisplay() );
-	}
-
-	/**
-	 * WordPress derives post_type from these screen IDs on its own, verified against core, so the
-	 * expected value is asserted rather than forced.
-	 */
-	private function set_screen( string $screen_id, string $post_type ): void {
-		set_current_screen( $screen_id );
-
-		$this->assertSame( $post_type, get_current_screen()->post_type );
 	}
 
 	/**
